Split functions into a new file. Created a first run file to ensure defaults are created.
